+++ R/utils_expand_data_to_grid.R
@@ -8,7 +8,7 @@
 #' @importFrom tidyselect ends_with
 # fac.typical indicates if factors should be held constant or not
 # need to be false for computing std.error for merMod objects
-get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, type = "fe", prettify = TRUE, prettify.at = 25, pretty.message = TRUE, condition = NULL) {
+get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, type = "fe", pretty.message = TRUE, condition = NULL) {
   # special handling for coxph
   if (inherits(model, "coxph")) mf <- dplyr::select(mf, -1)
 
@@ -63,9 +63,28 @@ get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, typ
     finally = function(x) { NULL }
   )
 
+
   # create unique combinations
-  rest <- rest[!(rest %in% names(first))]
-  first <- c(first, lapply(mf[, rest], function(i) sort(unique(i, na.rm = TRUE))))
+  xl.remain <- which(!(rest %in% names(first)))
+
+  # fix remaining x-levels
+  xl <- purrr::map(xl.remain, function(.x) {
+    pr <- mf[[rest[.x]]]
+    if (is.numeric(pr)) {
+      if (.x > 1 && dplyr::n_distinct(pr, na.rm = TRUE) >= 10)
+        rprs_values(pr)
+      else if (dplyr::n_distinct(pr, na.rm = TRUE) < 9)
+        na.omit(unique(pr))
+      else
+        pretty_range(pr)
+    } else if (is.factor(pr))
+      levels(pr)
+    else
+      na.omit(unique(pr))
+  })
+
+  names(xl) <- rest[xl.remain]
+  first <- c(first, xl)
 
   # get names of all predictor variable
   alle <- sjstats::pred_vars(model)
@@ -144,21 +163,6 @@ get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, typ
 
   # add constant values.
   first <- c(first, const.values)
-
-  # reduce and prettify elements with too many values
-  if (prettify) {
-    .pred <- function(p) is.numeric(p) && dplyr::n_distinct(p) > prettify.at
-    too.many <- purrr::map_lgl(first, .pred)
-    first <- purrr::map_if(first, .p = .pred, pretty_range)
-
-    if (any(too.many) && pretty.message) {
-      message(sprintf(
-        "Following variables had many unique values and were prettified: %s. Use `pretty = FALSE` to get smoother plots with all values, however, at the cost of increased memory usage.",
-        paste(names(first)[too.many], collapse = ", "))
-      )
-    }
-  }
-
 
   # create data frame with all unqiue combinations
   dat <- tibble::as_tibble(expand.grid(first))