+++ NEWS.md
@@ -2,8 +2,10 @@
 
 ## Changes to functions
 
+* `ggeffects()` and `ggpredict()` now automatically set numeric vector with 10 or more unique values to representative values (see `rprs_values()`), if these are used as second or third value in the `terms`-argument (to represent a grouping structure).
 * Fix memory allocation issue in `ggeffects()`.
 * `rprs_values()` is now exported.
+* The `pretty`-argument is deprecated, because prettifying values almost always makes sense - so this is done automatically.
 
 # ggeffects 0.4.0